requirements.txt
backcall==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.7
cycler==0.11.0
decorator==5.1.0
easydict==1.9
einops==0.3.2
future==0.18.2
graphviz==0.8.4
idna==3.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.1.0
kiwisolver==1.3.1
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mxnet-cu101==1.6.0.post0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603487797006/work
olefile==0.46
opencv-python==4.5.4.58
pandas==1.1.5
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1625649052827/work
prompt-toolkit==3.0.21
ptyprocess==0.7.0
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
scikit-learn==0.24.2
scipy==1.5.4
six @ file:///tmp/build/80754af9/six_1623709665295/work
thop==0.0.31.post2005241907
threadpoolctl==3.0.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.62.3
traitlets==4.3.3
urllib3==1.26.7
wcwidth==0.2.5
